+++ man-roxygen/yaml_generic.R
@@ -29,10 +29,13 @@
 #' If the author doesn't have an ORCID, you can omit that line.
 #' However, note that the INBO policy is that all scientific personnel is
 #' required to obtain an ORCID from https://orcid.org.
+#' Add `corresponding: true` to the corresponding author.
 #'
 #' ```
-#'   - firstname: "Authors first name"
-#'     name: "Authors last name"
+#'   - name:
+#'      given: "Authors first name"
+#'      family: "Authors last name"
 #'     email: "local-part@domain.com"
 #'     orcid: 0000-0002-1825-0097
+#'     corresponding: true
 #' ```